pyproject.toml
[project]
name = "marcura"
version = "0.1.0"
description = "Port Tariff Calculation System - Agentic RAG"
requires-python = ">=3.13"
dependencies = [
    "beartype>=0.22.9",
    "chromadb>=1.5.7",
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",

    "langchain-openai>=0.3.0",
    "openai>=1.75.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.13.0",

    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pyright>=1.1.408",
    "pytest>=8.0.0",
]

[tool.mypy]
strict = true
python_version = "3.13"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.13"
